requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
clickhouse-connect==0.9.1
clickhouse-driver==0.2.9
Flask==3.1.2
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
html5lib==1.1
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
lz4==4.4.5
MarkupSafe==3.0.3
MouseInfo==0.1.3
numpy==2.3.4
oauth2client==4.1.3
oauthlib==3.3.1
pandas==2.3.3
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-Quartz==12.0
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
rubicon-objc==0.5.2
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wikipedia==1.4.0
zstandard==0.25.0
